+++ src/test/java/edu/snu/vortex/compiler/optimizer/pass/ScheduleGroupPassTest.java
@@ -54,7 +54,7 @@ public final class ScheduleGroupPassTest {
     Integer previousScheduleGroupIndex = 0;
     for (final IRVertex irVertex : processedDAG.getTopologicalSort()) {
       assertTrue(irVertex.get(ExecutionProperty.Key.ScheduleGroupIndex) != null);
-      final Integer currentScheduleGroupIndex = (Integer) irVertex.get(ExecutionProperty.Key.ScheduleGroupIndex);
+      final Integer currentScheduleGroupIndex = irVertex.get(ExecutionProperty.Key.ScheduleGroupIndex);
       assertTrue(currentScheduleGroupIndex >= previousScheduleGroupIndex);
       if (currentScheduleGroupIndex > previousScheduleGroupIndex) {
         previousScheduleGroupIndex = currentScheduleGroupIndex;
